+++ changelog.md
@@ -1,4 +1,10 @@
 # Changelog
 
-**./submission/coding/1-reverse_string.md**
-* Revision complete.
+**./submission/coding/3-time_string.md**
+* Merged 'hoursBase' into 'hoursRounded'
+* Replaced `console.log(TimeConvert(readline()));` with working example.
+
+---
+
+**./submission/coding/work-file.js**
+* Used to revise question 3.